% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earth.R
\name{plot_satellite_image_and_pallette}
\alias{plot_satellite_image_and_pallette}
\title{plot_satellite_image_and_pallette}
\usage{
plot_satellite_image_and_pallette(latitude = 24.2, longitude = -77.88,
  zoom = 11, number_of_colors = 5)
}
\arguments{
\item{latitude}{}

\item{longitude}{}

\item{zoom}{}

\item{number_of_colors}{}
}
\description{
This returns a color scheme from a geographic place but 
as a side product, it plots a map of the satellite image along with the color pallette.
}
\examples{
#
#
uluru<-plot_satellite_image_and_pallette(latitude = -25.5,
longitude = 131,zoom=10)


world<-plot_satellite_image_and_pallette(latitude = 0,longitude = 0,
zoom=2,number_of_colors=8)

british_columbia_glacier<-plot_satellite_image_and_pallette(latitude = 50.759,
longitude = -125.673,zoom=10,number_of_colors=4)

 joshua_tree<-plot_satellite_image_and_pallette(latitude = 33.9,
 longitude = -115.9,zoom=9,number_of_colors=5)

bahamas<-plot_satellite_image_and_pallette(latitude = 24.2,
longitude=-77.88,zoom=11,number_of_colors=5)


rio_negro_amazon_mixing<-plot_satellite_image_and_pallette(latitude = -3.0817,
longitude=-60.49666,zoom=10,number_of_colors=3)

grand_canyon<-plot_satellite_image_and_pallette(latitude = 36.094994,
longitude=-111.837962,zoom=12,number_of_colors=6)

}

