% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earth.R
\name{print.palette}
\alias{print.palette}
\title{Print Method for Palette Objects}
\usage{
\method{print}{palette}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"palette"}.}

\item{...}{Additional arguments passed to plotting methods.}
}
\value{
No return value; called for its side effect of plotting.
}
\description{
Visualizes the palette and associated satellite image.
}
