% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earth.R
\name{get_earthtones}
\alias{get_earthtones}
\title{earthtones}
\usage{
get_earthtones(latitude = 50.759, longitude = -125.673, zoom = 11,
  number_of_colors = 3, method = "kmeans", sampleRate = 500)
}
\arguments{
\item{latitude}{}

\item{longitude}{}

\item{zoom}{}

\item{number_of_colors}{}

\item{method}{specifies clustering method. Options are \code{kmeans} or \code{pam} (partitioning around medoids)}

\item{sampleRate}{subsampling factor - bigger number = more subsampling}
}
\description{
Getting a color scheme from a place on earth
}
\examples{
#
#
uluru<-get_earthtones(latitude = -25.5,
longitude = 131,zoom=12,number_of_colors=5)
print(uluru)
plot(uluru)

\dontrun{
world<-get_earthtones(latitude = 0,longitude = 0,
zoom=2,number_of_colors=8)
plot(world)

british_columbia_glacier<-get_earthtones(latitude = 50.759
 ,longitude = -125.673,zoom=10,number_of_colors=5)
plot(british_columbia_glacier)

joshua_tree<-get_earthtones(latitude = 33.9, 
 longitude = -115.9,zoom=9,number_of_colors=5)
plot(joshua_tree)
 
## Compare clustering methods
par(mfrow=c(2,1))
bahamas<-get_earthtones(latitude = 24.2,longitude=-77.88,
zoom=11,number_of_colors=5,sampleRate=500)
plot(bahamas)

bahamas<-get_earthtones(latitude = 24.2,longitude=-77.88,
zoom=11,number_of_colors=5,method='pam',sampleRate=500)
plot(bahamas)
}



}

